go.mod: set go version to 1.17
This is not a new requirement... I started using 1.17 features at some
point in the past (as indicated in the CI tests).  I just missed
declaring that requirement in the go.mod file.

With go1.17, the structure of the go.mod files changes a bit, but this
shouldn't actually have any actual changes in the dependency tree.